requirements.txt
pytorchyolo
opencv-python
tqdm
pyyaml
pandas
seaborn
natsort
kornia
setuptools
numpy
pillow
transformers
imgaug
terminaltables
torchsummary
packaging
tensorboardx
requests
cython
datasets
six
scipy
jupyter
notebook
timm